refactor: move the organization switch logic to the component, so that it can be reused in the navbar
