+++ apps/web/src/pages/customer.tsx
@@ -1,47 +1,17 @@
-import { useToast } from "@/hooks/toast/useToast";
 import LoadingSpinner from "@/ui/LoadingSpinner";
 import OrganizationSwitcher from "@/ui/OrganizationSwitcher";
 import H1 from "@/ui/typography/H1";
 import H2 from "@/ui/typography/H2";
 import Subtle from "@/ui/typography/Subtle";
-import { trpc } from "@/utils/trpc";
 import { userIsMemberOfAnyOrganization } from "@acme/api/src/lib/clerk";
 import { useAuth } from "@clerk/nextjs";
 import { getAuth } from "@clerk/nextjs/server";
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
-import { useEffect } from "react";
 
 type CustomerPageProps = InferGetServerSidePropsType<typeof getServerSideProps>;
 
 const CustomerPage = ({ isMemberOfAnyOrganization }: CustomerPageProps) => {
-  const { userId, orgId, isLoaded } = useAuth();
-  const { toast } = useToast();
-  const { mutateAsync, isLoading } =
-    trpc.organization.addUserToOrganization.useMutation({
-      onSuccess: () => {
-        toast({
-          title: "Success",
-          description: "Successfully set active organization",
-          severity: "success",
-        });
-      },
-      onError: () => {
-        toast({
-          title: "Error",
-          description: "An error occurred when setting the active organization",
-          severity: "error",
-        });
-      },
-    });
-
-  useEffect(() => {
-    const addUserToOrganization = async () => {
-      if (!orgId) return;
-      await mutateAsync({ userId });
-    };
-
-    addUserToOrganization();
-  }, [mutateAsync, orgId, userId]);
+  const { isLoaded } = useAuth();
 
   return (
     <div>
@@ -49,13 +19,6 @@ const CustomerPage = ({ isMemberOfAnyOrganization }: CustomerPageProps) => {
       <Subtle>
         Before you can use the platform, you need to set an active organization.
       </Subtle>
-
-      {isLoading && (
-        <div className="flex items-center justify-center gap-x-2">
-          <LoadingSpinner />
-          <Subtle>Setting active organization</Subtle>
-        </div>
-      )}
 
       {!isMemberOfAnyOrganization && (
         <div>
